shortcut plotlist plot construction/initialization
